+++ platform/darwin/src/MGLMultiPoint_Private.h
@@ -13,6 +13,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @class MGLPolygon;
 @class MGLPolyline;
+@class MGLCircle;
 
 @protocol MGLMultiPointDelegate;
 
@@ -36,7 +37,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (mbgl::Color)strokeColorForShapeAnnotation:(MGLShape *)annotation;
 
 /** Returns the fill color object for the given annotation. */
-- (mbgl::Color)fillColorForPolygonAnnotation:(MGLPolygon *)annotation;
+- (mbgl::Color)fillColorForShape:(MGLShape *)annotation;
 
 /** Returns the stroke width object for the given annotation. */
 - (CGFloat)lineWidthForPolylineAnnotation:(MGLPolyline *)annotation;