+++ platform/macos/src/Mapbox.h
@@ -32,6 +32,7 @@ FOUNDATION_EXPORT MGL_EXPORT const unsigned char MapboxVersionString[];
 #import "MGLPointCollection.h"
 #import "MGLPolygon.h"
 #import "MGLPolyline.h"
+#import "MGLCircle.h"
 #import "MGLShape.h"
 #import "MGLShapeCollection.h"
 #import "MGLStyle.h"